stress_client/requirements.txt
Django==4.1.5
json2html==1.3.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.1
sectionproperties==2.1.5
psycopg2
pytz
tzdata
openpyxl

numpy==1.21.2
scipy==1.9.3

matplotlib==3.7.0

pandas==1.5.2

gunicorn==20.0.4

backports.zoneinfo==0.2.1